requirements.txt
PySide6>=6.6
numpy>=1.24
Pillow>=10.0
pandas>=1.5
pyarrow>=11.0
